refactor(library): complete the virtual window behind one geometry interface
Extend the Library Browser window module to own virtual geometry:
shared-viewport measurement, enablement, overscan, canvas height, and
row placement, while the route becomes a rendering adapter over the
module's virtual interface. The virtualizer still activates only after
Solid mount, virtual mode, and a real shared viewport; overscan stays
adaptive within 6-18 rows with a six-row fallback for invalid geometry;
the canvas keeps full height with scroll-margin row placement and no
entrance fade on the virtual root. The native offset observer and the
no-blank-frame contract are preserved, proven by the shell blanking
tests and the focused native WebView virtual-scroll spec. The
ADR-governed Library Image Cache is untouched.

Closes #208
